Reorganize docstring parsing and Markdown modules
