+++ crates/ty_ide/src/docstring/document/preformatted.rs
@@ -1,12 +1,49 @@
 use ruff_python_trivia::leading_indentation;
 use ruff_text_size::TextSize;
 
-use super::markdown;
+/// Represents a fenced Markdown code block.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub(in crate::docstring) struct MarkdownFence<'a> {
+    /// The string used to denote the start and end of the fenced code block.
+    marker: &'a str,
+}
+
+impl<'a> MarkdownFence<'a> {
+    pub(in crate::docstring) fn marker(&self) -> &'a str {
+        self.marker
+    }
+
+    /// Recognizes the beginning of a fenced code block if one is present on the given line.
+    pub(in crate::docstring) fn find(line: &'a str) -> Option<Self> {
+        let line = line.trim_start_matches(' ');
+        let has_tick_fence = line.starts_with("```");
+        let has_tilde_fence = line.starts_with("~~~");
+        if !has_tick_fence && !has_tilde_fence {
+            return None;
+        }
+
+        let without_leading_fence = if has_tick_fence {
+            line.trim_start_matches('`')
+        } else {
+            line.trim_start_matches('~')
+        };
+        let fence_len = line.len() - without_leading_fence.len();
+        let fence = &line[..fence_len];
+
+        // We *don't* want to consider ```hello``` as a codefence; that's inline code!
+        (!without_leading_fence.contains(fence)).then_some(Self { marker: fence })
+    }
+
+    /// Returns whether `line` closes this fenced code block.
+    pub(in crate::docstring) fn is_closed_by(&self, line: &str) -> bool {
+        line.trim_start_matches(' ').starts_with(self.marker)
+    }
+}
 
 /// Recognizes preformatted blocks that may occur within a docstring.
 #[derive(Debug, Default, Clone, PartialEq, Eq)]
 pub(super) struct PreformattedBlockScanner<'a> {
-    active_markdown_fence: Option<markdown::MarkdownFence<'a>>,
+    active_markdown_fence: Option<MarkdownFence<'a>>,
     active_doctest: bool,
     rest_literal_blocks: RestLiteralBlockScanner,
 }
@@ -43,7 +80,7 @@ impl<'a> PreformattedBlockScanner<'a> {
             return true;
         }
 
-        if let Some(fence) = markdown::MarkdownFence::find(line) {
+        if let Some(fence) = MarkdownFence::find(line) {
             self.active_markdown_fence = Some(fence);
             return true;
         }
